feat: create SignaturesClient component for managing TCC signatures
